utils: extract cache
